capstone/voice to nlp/requirements.txt
google-generativeai
pyaudio
whisper
numpy
pyttsx3
SpeechRecognition
sounddevice
scipy
gtts